+++ src/Symfony/Component/Mailer/Bridge/Amazon/Tests/Transport/SesTransportFactoryTest.php
@@ -28,7 +28,7 @@ class SesTransportFactoryTest extends TransportFactoryTestCase
         return new SesTransportFactory($this->getDispatcher(), $this->getClient(), $this->getLogger());
     }
 
-    public function supportsProvider(): iterable
+    public static function supportsProvider(): iterable
     {
         yield [
             new Dsn('ses+api', 'default'),
@@ -61,7 +61,7 @@ class SesTransportFactoryTest extends TransportFactoryTestCase
         ];
     }
 
-    public function createProvider(): iterable
+    public static function createProvider(): iterable
     {
         $client = $this->getClient();
         $dispatcher = $this->getDispatcher();
@@ -158,7 +158,7 @@ class SesTransportFactoryTest extends TransportFactoryTestCase
         ];
     }
 
-    public function unsupportedSchemeProvider(): iterable
+    public static function unsupportedSchemeProvider(): iterable
     {
         yield [
             new Dsn('ses+foo', 'default', self::USER, self::PASSWORD),
@@ -166,7 +166,7 @@ class SesTransportFactoryTest extends TransportFactoryTestCase
         ];
     }
 
-    public function incompleteDsnProvider(): iterable
+    public static function incompleteDsnProvider(): iterable
     {
         yield [new Dsn('ses+smtp', 'default', self::USER)];
     }