+++ src/Symfony/Component/Mailer/Tests/Transport/Smtp/EsmtpTransportFactoryTest.php
@@ -25,7 +25,7 @@ class EsmtpTransportFactoryTest extends TransportFactoryTestCase
         return new EsmtpTransportFactory($this->getDispatcher(), $this->getClient(), $this->getLogger());
     }
 
-    public function supportsProvider(): iterable
+    public static function supportsProvider(): iterable
     {
         yield [
             new Dsn('smtp', 'example.com'),
@@ -43,7 +43,7 @@ class EsmtpTransportFactoryTest extends TransportFactoryTestCase
         ];
     }
 
-    public function createProvider(): iterable
+    public static function createProvider(): iterable
     {
         $eventDispatcher = $this->getDispatcher();
         $logger = $this->getLogger();