+++ src/Symfony/Component/Notifier/Bridge/MessageMedia/Tests/MessageMediaTransportFactoryTest.php
@@ -25,7 +25,7 @@ final class MessageMediaTransportFactoryTest extends TransportFactoryTestCase
         return new MessageMediaTransportFactory();
     }
 
-    public function createProvider(): iterable
+    public static function createProvider(): iterable
     {
         yield [
             'messagemedia://host.test',
@@ -38,13 +38,13 @@ final class MessageMediaTransportFactoryTest extends TransportFactoryTestCase
         ];
     }
 
-    public function supportsProvider(): iterable
+    public static function supportsProvider(): iterable
     {
         yield [true, 'messagemedia://apiKey:apiSecret@default'];
         yield [false, 'somethingElse://apiKey:apiSecret@default'];
     }
 
-    public function unsupportedSchemeProvider(): iterable
+    public static function unsupportedSchemeProvider(): iterable
     {
         yield ['somethingElse://apiKey:apiSecret@default'];
     }