+++ src/Symfony/Component/Notifier/Bridge/OvhCloud/Tests/OvhCloudTransportFactoryTest.php
@@ -25,7 +25,7 @@ final class OvhCloudTransportFactoryTest extends TransportFactoryTestCase
         return new OvhCloudTransportFactory();
     }
 
-    public function createProvider(): iterable
+    public static function createProvider(): iterable
     {
         yield [
             'ovhcloud://host.test?consumer_key=consumerKey&service_name=serviceName',
@@ -38,19 +38,19 @@ final class OvhCloudTransportFactoryTest extends TransportFactoryTestCase
         ];
     }
 
-    public function supportsProvider(): iterable
+    public static function supportsProvider(): iterable
     {
         yield [true, 'ovhcloud://key:secret@default?consumer_key=consumerKey&service_name=serviceName&sender=sender'];
         yield [false, 'somethingElse://key:secret@default?consumer_key=consumerKey&service_name=serviceName&sender=sender'];
     }
 
-    public function missingRequiredOptionProvider(): iterable
+    public static function missingRequiredOptionProvider(): iterable
     {
         yield 'missing option: consumer_key' => ['ovhcloud://key:secret@default?service_name=serviceName'];
         yield 'missing option: service_name' => ['ovhcloud://key:secret@default?consumer_key=consumerKey'];
     }
 
-    public function unsupportedSchemeProvider(): iterable
+    public static function unsupportedSchemeProvider(): iterable
     {
         yield ['somethingElse://key:secret@default?consumer_key=consumerKey&service_name=serviceName&sender=sender'];
         yield ['somethingElse://key:secret@default?service_name=serviceName'];