+++ src/Symfony/Component/Translation/Bridge/Crowdin/Tests/CrowdinProviderTest.php
@@ -33,7 +33,7 @@ class CrowdinProviderTest extends ProviderTestCase
         return new CrowdinProvider($client, $loader, $logger, $this->getXliffFileDumper(), $defaultLocale, $endpoint);
     }
 
-    public function toStringProvider(): iterable
+    public static function toStringProvider(): iterable
     {
         yield [
             $this->createProvider($this->getClient()->withOptions([
@@ -552,7 +552,7 @@ XLIFF;
         $provider->write($translatorBag);
     }
 
-    public function getResponsesForProcessAddFileAndUploadTranslations(): \Generator
+    public static function getResponsesForProcessAddFileAndUploadTranslations(): \Generator
     {
         $arrayLoader = new ArrayLoader();
 
@@ -660,7 +660,7 @@ XLIFF
         $this->assertEquals($expectedTranslatorBag->getCatalogues(), $translatorBag->getCatalogues());
     }
 
-    public function getResponsesForOneLocaleAndOneDomain(): \Generator
+    public static function getResponsesForOneLocaleAndOneDomain(): \Generator
     {
         $arrayLoader = new ArrayLoader();
 
@@ -773,7 +773,7 @@ XLIFF
         $this->assertEquals($expectedTranslatorBag->getCatalogues(), $translatorBag->getCatalogues());
     }
 
-    public function getResponsesForDefaultLocaleAndOneDomain(): \Generator
+    public static function getResponsesForDefaultLocaleAndOneDomain(): \Generator
     {
         $arrayLoader = new ArrayLoader();