+++ src/Symfony/Component/Mailer/CHANGELOG.md
@@ -1,6 +1,14 @@
 CHANGELOG
 =========
 
+5.4.21
+------
+
+* [BC BREAK] The following data providers for `TransportFactoryTestCase` are now static:
+  `supportsProvider()`, `createProvider()`, `unsupportedSchemeProvider()`and `incompleteDsnProvider()`
+* [BC BREAK] The following data providers for `TransportTestCase` are now static:
+  `toStringProvider()`, `supportedMessagesProvider()` and `unsupportedMessagesProvider()`
+
 5.4
 ---